Parsing into "DB"
It seems to work, the parts have been tested apart from each other and they work together with a result that looks correct. It is still uncertain because it hasn't been checked. The old parsing function is still in server.js commented away just in case, and the new system needs to be implemented to the server with a function for searching a food shift.