Exam info system added
It works only based on the current date. A better system would require good date functions, and right now if I want to make somethng for this project I should focus on either getting it working or making the code less messy (aka writing a big part of it again...). This was my first project of this scale, and the first one that seems to get finished & even published.